requirements.txt
bcrypt==4.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dash==2.18.2
dash-ag-grid==31.2.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flask==3.0.3
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
plotly==5.24.1
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
SQLAlchemy==2.0.36
tenacity==9.0.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.21.0
